generator/requirements.txt
# Markdown Parsing
markdown-it-py

# Templating
Jinja2

# Database (should match admin_app choice)
# pymongo
motor # If generator needs to be async

# Environment variable loading (optional)
# python-dotenv 